+++ xla/pjrt/common_pjrt_client.h
@@ -754,6 +754,9 @@ class CommonPjRtBufferImpl : public CommonPjRtBuffer {
   Future<> CopyRawToHost(void* dst, int64_t offset,
                          int64_t transfer_size) override;
 
+  Future<> CopyRawFromHost(const void* src, int64_t offset,
+                           int64_t transfer_size) override;
+
   Future<> CopyRawToHostFuture(Future<void*> dst, int64_t offset,
                                int64_t transfer_size) override;